Phase F-1 wave 1: scaffold _ops/ + extract _StoreCore (#114)
Stand up the mixin-split inheritance scaffold without moving any
per-resource ops yet:

- New `_ops/_helpers.py` carries the self-free module helpers
  (`_validated_update`, `_deep`, `_no_op_*`, `_resolve_trees`, the two
  message builders). `_base.py` re-exports them at module scope so the
  eden-git integrator tests' `from eden_storage._base import
  _validated_update` keeps working.
- New `_StoreCore` carries `__init__`, the event-id factory, the
  `_event`/`_ts`/`_maybe_ts` helpers, the cross-resource read-side
  predicates (`_require_*`, `_find_starting_variant_for_implement_task`,
  `_validate_registry_id`), and every backend-primitive declaration
  (per plan §D.4/§D.5). `_reseed_default_event_counter`'s `__func__`
  comparison now references `_StoreCore._default_event_id`.
- `_StoreBase(_StoreCore)`: every per-resource op stays inline for now;
  mixin extraction is waves 2-4.

Behavior-preserving: storage suite 373 passed / 185 skipped (postgres
auto-skips locally), full pyright clean, eden-git + checkpoint
downstream importers green. `_base.py` still > 800 SLOC; the
`# slop-allow-file:` annotation stays until wave 5.
