+++ reference/packages/eden-storage/src/eden_storage/_ops/__init__.py
@@ -1,0 +1,8 @@
+"""Per-resource operation mixins composed into ``_StoreBase``.
+
+See [`.._base`](../_base.py) for ``_StoreCore`` (the abstract core every
+mixin inherits) and the composite ``_StoreBase`` that flattens the
+mixin MRO. The refactor that introduced this package is recorded in
+[`docs/plans/refactor-f1-storebase-split.md`](../../../../../../docs/plans/refactor-f1-storebase-split.md)
+(issue #114).
+"""
